+++ test/slugger.test.ts
@@ -675,5 +675,20 @@ describe('slugger', () => {
         expect(utils.isModel(Model)).toEqual(true);
       });
     });
+
+    describe('edge cases', () => {
+      it('no valid slug characters - fall back to _id', () => {
+        const doc = new Model({ firstname: ':-)', lastname: 'doe' });
+        const generator = utils.createDefaultGenerator('firstname');
+        expect(generator(doc, 0)).toEqual(String(doc._id));
+      });
+
+      it('no valid slug characters - fall back to _id in combination with maxLength', () => {
+        const doc = new Model({ firstname: ':-)', lastname: 'doe' });
+        const generator = utils.createDefaultGenerator('firstname');
+        const docId = String(doc._id);
+        expect(generator(doc, 0, 10)).toEqual(docId.substring(docId.length - 10));
+      });
+    });
   });
 });